requirements.txt
functions-framework==3.*
Werkzeug>=3.0.0

numpy>=1.26.0
pandas>=2.0.0
scikit-learn>=1.3.0
 

joblib>=1.3.0
google-cloud-storage>=2.0.0
